feat(user): kepala sekolah

+++ be/src/controllers/VerifikatorController.js
@@ -13,6 +13,11 @@ class VerifikatorController {
 				return res.status(400).json({ message: "Nama, username, dan password wajib diisi" });
 			}
 
+			const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z\d]).{6,}$/;
+			if (!passwordRegex.test(password)) {
+				return res.status(400).json({ message: "Kata sandi minimal 6 karakter dan harus mengandung huruf kecil, huruf besar, angka, dan simbol." });
+			}
+
 			// Cek duplikasi
 			const existingVerifikator = await VerifikatorService.getVerifikatorByUsername(username);
 			if (existingVerifikator) {
@@ -92,6 +97,15 @@ class VerifikatorController {
 	updateVerifikatorData = async (req, res) => {
 		try {
 			const { id } = req.params;
+			const { password } = req.body;
+
+			if (password && password.trim() !== "") {
+				const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z\d]).{6,}$/;
+				if (!passwordRegex.test(password)) {
+					return res.status(400).json({ message: "Kata sandi minimal 6 karakter dan harus mengandung huruf kecil, huruf besar, angka, dan simbol." });
+				}
+			}
+
 			const updated = await VerifikatorService.updateVerifikator(id, req.body);
 			res.status(200).json({ message: "Update verifikator berhasil", data: updated });
 		} catch (error) {
@@ -103,6 +117,11 @@ class VerifikatorController {
 		try {
 			const { id } = req.params;
 			const { status_aktif } = req.body;
+
+			if (typeof status_aktif !== "boolean") {
+				return res.status(400).json({ message: "status_aktif harus berupa boolean" });
+			}
+
 			const updated = await VerifikatorService.updateStatusVerifikator(id, status_aktif);
 			res.status(200).json({ message: "Status verifikator berhasil diperbarui", data: updated });
 		} catch (error) {
@@ -113,6 +132,16 @@ class VerifikatorController {
 	remove = async (req, res) => {
 		try {
 			const { id } = req.params;
+			const verifikator = await VerifikatorService.getVerifikatorById(id);
+			
+			if (!verifikator) {
+				return res.status(404).json({ message: "Verifikator tidak ditemukan" });
+			}
+			
+			if (verifikator.status_aktif) {
+				return res.status(400).json({ message: "Akun aktif tidak dapat dihapus." });
+			}
+
 			await VerifikatorService.deleteVerifikator(id);
 			res.status(200).json({ message: "Hapus verifikator berhasil" });
 		} catch (error) {
